# perf(compile): read embedded files as static references when UTF-8 and reading as strings