--- a/cli/util/text_encoding.rs
+++ b/cli/util/text_encoding.rs
@@ -12,7 +12,7 @@ static SOURCE_MAP_PREFIX: &[u8] =
   b"//# sourceMappingURL=data:application/json;base64,";
 
 #[inline(always)]
-pub fn from_utf8_lossy_cow<'a>(bytes: Cow<'a, [u8]>) -> Cow<'a, str> {
+pub fn from_utf8_lossy_cow(bytes: Cow<[u8]>) -> Cow<str> {
   match bytes {
     Cow::Borrowed(bytes) => String::from_utf8_lossy(bytes),
     Cow::Owned(bytes) => Cow::Owned(from_utf8_lossy_owned(bytes)),