--- a/resolvers/deno/fs.rs
+++ b/resolvers/deno/fs.rs
@@ -1,5 +1,6 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 
+use std::borrow::Cow;
 use std::path::Path;
 use std::path::PathBuf;
 
@@ -10,7 +11,10 @@ pub struct DirEntry {
 }
 
 pub trait DenoResolverFs {
-  fn read_to_string_lossy(&self, path: &Path) -> std::io::Result<String>;
+  fn read_to_string_lossy(
+    &self,
+    path: &Path,
+  ) -> std::io::Result<Cow<'static, str>>;
   fn realpath_sync(&self, path: &Path) -> std::io::Result<PathBuf>;
   fn exists_sync(&self, path: &Path) -> bool;
   fn is_dir_sync(&self, path: &Path) -> bool;