**Speed**

```ts
let count = 0;
for (var i = 0; i < 500; i++) {
  const text = Deno.readTextFileSync(import.meta.dirname + "/test.txt");
  count += text.length;
}

console.log(count);
```

```shellsession
% hyperfine --warmup=4 './old' './new'
Benchmark 1: ./old
  Time (mean ± σ):      2.236 s ±  0.009 s    [User: 2.164 s, System: 0.423 s]
  Range (min … max):    2.224 s …  2.250 s    10 runs
 
Benchmark 2: ./new
  Time (mean ± σ):      1.570 s ±  0.003 s    [User: 1.552 s, System: 0.013 s]
  Range (min … max):    1.566 s …  1.574 s    10 runs
 
Summary
  ./new ran
    1.42 ± 0.01 times faster than ./old
```

I also checked perf of just reading a file from the file system and it is the same as before (expected... just was making sure that didn't regress somehow)

**Memory Usage**

```js
import process from "node:process";

const data = [];
setInterval(() => {
  const text = Deno.readTextFileSync(import.meta.dirname + "/test.txt");
  data.push(text);
  console.log(process.memoryUsage().rss);
}, 1000);
```

```shellsession
scratch2 % ./old     
54738944
61014016
73236480
85245952
97648640
109740032
122765312
^C
scratch2 % ./new     
43417600
43548672
43859968
43876352
43876352
44154880
44171264
^C
```